Fixed camera directory display and html title

+++ app/models/camera.rb
@@ -13,22 +13,23 @@ class Camera
     directory = File.join( Rails.root , '../okaapi', 'public', 'camera')
     if Dir.exists? directory    
       
-      datelist = Dir.entries(directory).reject{|entry| entry =~ /^\.{1,2}$/}.sort_by { |a| File.stat(File.join(directory,a)).mtime }.reverse
+      datelist = Dir.entries(directory).reject{|entry| entry =~ /^\.{1,2}$/}.sort #_by { |a| File.stat(File.join(directory,a)).mtime }.reverse
       
       datelist.each do |date|
-        out << "<h3> #{Camera.degarble(date)} <small><a href='/camera_directory_delete/#{filter}/#{date}/#{page}' " 
-        out <<     " onclick='return confirm(\"Are you sure?\")'> delete </a></small> </h3>"                 
         datedirectory = File.join( directory, date )  
-              
         imagefiles = Dir.entries(datedirectory).sort 
-        imagefiles.each do |imagefile|
-          if imagefile.include? filter
-            t = imagefile.gsub(/#{filter}/,'').gsub(/.jpg/,'')
-            ##out << "<a href='/camera/#{date}/#{imagefile}'>"
-            out << "<img src='/camera/#{date}/#{imagefile}' width='64' style='border:2px solid white' title='#{t}' "
-            out << " onclick=\"this.style.width='64px'\" ondblclick=\"this.style.width='500px'\" >\n"
+        if imagefiles.find{|each| each.include? filter}
+          out << "<h3> #{Camera.degarble(date)} <small><a href='/camera_directory_delete/#{filter}/#{date}/#{page}' " 
+          out <<     " onclick='return confirm(\"Are you sure?\")'> delete </a></small> </h3>"                 
+          imagefiles.each do |imagefile|
+            if imagefile.include? filter
+              t = imagefile.gsub(/#{filter}/,'').gsub(/.jpg/,'')
+              ##out << "<a href='/camera/#{date}/#{imagefile}'>"
+              out << "<img src='/camera/#{date}/#{imagefile}' width='64' style='border:2px solid white' title='#{t}' "
+              out << " onclick=\"this.style.width='64px'\" ondblclick=\"this.style.width='500px'\" >\n"
             
-            ##out << "</a>"
+              ##out << "</a>"
+            end
           end
         end
